# internal/api/app/swagger.yml
openapi: 3.0.3
info:
  title: MockBank App
  description: BFF for the MockBank application.
  version: 1.0.0
paths:
...
  /api/directory/callback:
    get:
      operationId: handleDirectoryCallback
      security:
        - SessionCookie: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '303':
          description: See Other - Redirect to the app
          headers:
            Location:
              $ref: '#/components/headers/Location'
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
  /api/me:
    get:
      summary: Get current user information
      operationId: getCurrentUser
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/logout:
    post:
      summary: Logout the current user
      operationId: logoutUser
      security:
        - SessionCookie: []
      responses:
        '303':
          description: See Other - Redirect to the login page
          headers:
            Location:
              $ref: '#/components/headers/Location'
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
  /api/orgs/{orgId}/users:
    get:
      summary: List mock users in an organization
      operationId: getMockUsers
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockUsersResponse'
    post:
      summary: Create a new user in an organization
      operationId: createMockUser
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockUserResponse'
  /api/orgs/{orgId}/users/{userId}:
    put:
      summary: Update a user in an organization
      operationId: updateMockUser
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockUserResponse'
    delete:
      summary: Delete a mock user in an organization
      operationId: deleteMockUser
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
      responses:
        '204':
          description: Mock user deleted
  /api/orgs/{orgId}/users/{userId}/businesses/{businessId}/bind:
    post:
      summary: Bind a user to a business
      operationId: bindUserToBusiness
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
        - $ref: '#/components/parameters/businessId'
      responses:
        '201':
          description: User successfully bound to business
  /api/orgs/{orgId}/users/{userId}/accounts:
    post:
      summary: Create a new account
      operationId: createAccount
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
    get:
      summary: Get accounts of a user
      operationId: getAccounts
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: User accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsResponse'
  /api/orgs/{orgId}/users/{userId}/accounts/{accountId}:
    put:
      summary: Update an account
      operationId: updateAccount
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
        - $ref: '#/components/parameters/accountId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
    delete:
      summary: Delete an account
      operationId: deleteAccount
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
        - $ref: '#/components/parameters/accountId'
      responses:
        '204':
          description: Account deleted
  /api/orgs/{orgId}/users/{userId}/consents:
    get:
      summary: Get consents of a user
      operationId: getConsents
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: User consents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentsResponse'
  /api/orgs/{orgId}/users/{userId}/resources:
    get:
      summary: Get shared resources of a user
      operationId: getResources
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: User resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesResponse'
  /api/orgs/{orgId}/users/{userId}/resources/{resourceId}/consents/{consentId}:
    patch:
      summary: Update status of a shared resource
      operationId: patchResourceStatus
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/organizationId'
        - $ref: '#/components/parameters/mockUserId'
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
        - name: consentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ResourceType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
              properties:
                data: 
                  type: object
                  required:
                    - status
                  properties:
                    status:
                      $ref: '#/components/schemas/ResourceStatus'
      responses:
        '204':
          description: Status updated successfully
components:
  schemas:
    UserResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - username
            - organizations
          properties:
            username:
              type: string
            organizations:
              type: array
              items:
                type: object
                required:
                  - id
                  - name
                properties:
                  id:
                    type: string
                  name:
                    type: string
    MockUsersResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            type: object
            required:
              - id
              - name
              - username
              - cpf
            properties:
              id:
                type: string
              name:
                type: string
              username:
                type: string
              cpf:
                type: string
              description:
                type: string
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'
    MockUserRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - username
            - name
            - cpf
          properties:
            username:
              type: string
            password:
              type: string
            name:
              type: string
            cpf:
              type: string
            description:
              type: string
    MockUserResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - id
            - username
            - name
            - cpf
          properties:
            id:
              type: string
            username:
              type: string
            password:
              type: string
            name:
              type: string
            cpf:
              type: string
            description:
              type: string
...
    AccountRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - number
            - type
            - subtype
            - availableAmount
            - blockedAmount
            - automaticallyInvestedAmount
          properties:
            number:
              type: string
              maxLength: 20
            type:
              type: string
            subtype:
              type: string
            availableAmount:
              type: string
            blockedAmount:
              type: string
            automaticallyInvestedAmount:
              type: string
            overdraftLimitContracted:
              type: string
            overdraftLimitUsed:
              type: string
            overdraftLimitUnarraged:
              type: string
    AccountResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/AccountData"
    AccountsResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AccountData"
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'
...
    ConsentsResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            type: object
            required:
              - consentId
              - status
              - permissions
              - creationDateTime
              - statusUpdateDateTime
              - userId
              - clientId
            properties:
              consentId:
                type: string
              status:
                type: string
              permissions:
                type: array
                items:
                  type: string
              creationDateTime:
                type: string
                format: date-time
              statusUpdateDateTime:
                type: string
                format: date-time
              expirationDateTime:
                type: string
                format: date-time
              rejectedBy:
                type: string
              rejectionReason:
                type: string
              userId:
                type: string
              clientId:
                type: string
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'
    ResourcesResponse:
      type: object
      required:
        - data
        - links
        - meta
      properties:
        data:
          type: array
          minItems: 0
          items:
            type: object
            required:
              - resourceId
              - consentId
              - type
              - status
              - creationDateTime
            properties:
              resourceId:
                type: string
              consentId:
                type: string
              type:
                $ref: "#/components/schemas/ResourceType"
              status:
                $ref: "#/components/schemas/ResourceStatus"
              creationDateTime:
                type: string
                format: date-time
        links:
          $ref: '#/components/schemas/Links'
        meta:
          $ref: '#/components/schemas/Meta'
    ResourceStatus:
      type: string
      enum:
        - AVAILABLE
        - UNAVAILABLE
        - TEMPORARILY_UNAVAILABLE
        - PENDING_AUTHORISATION
    ResourceType:
      type: string
      enum:
        - ACCOUNT
        - CREDIT_CARD_ACCOUNT
        - LOAN
        - FINANCING
        - UNARRANGED_ACCOUNT_OVERDRAFT
        - INVOICE_FINANCING
        - BANK_FIXED_INCOME
        - CREDIT_FIXED_INCOME
        - VARIABLE_INCOME
        - TREASURE_TITLE
        - FUND
        - EXCHANGE
...
    Meta:
      type: object
      description: Meta informações referente à API requisitada.
      required:
        - totalRecords
        - totalPages
        - requestDateTime
      properties:
        totalRecords:
          type: integer
          format: int32
          description: Número total de registros no resultado
          example: 1
        totalPages:
          type: integer
          format: int32
          description: Número total de páginas no resultado
          example: 1
        requestDateTime:
          description: 'Data e hora da consulta, conforme especificação RFC-3339, formato UTC.'
          type: string
          maxLength: 20
          format: date
          example: '2021-05-21T08:30:00Z'
...
  parameters:
    organizationId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
    mockUserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    businessId:
      name: businessId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      description: Número da página que está sendo requisitada (o valor da primeira página é 1).
      schema:
        type: integer
        default: 1
        minimum: 1
        maximum: 2147483647
        format: int32
    pageSize:
      name: page-size
      in: query
      description: Quantidade total de registros por páginas.
      schema:
        type: integer
        default: 25
        minimum: 1
        format: int32
        maximum: 1000
  headers:
    Location:
      description: URL to which the client is redirected
      schema:
        type: string
    SessionCookie:
      description: Session ID for authentication
      schema:
        type: string
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session ID for authentication